+++ prisma/schema.prisma
@@ -15,7 +15,7 @@ model Brand {
   id String   @id @default(uuid())
   name String
   country String
-  year String
+  year Int
   headquarter String
   website String
   created_at DateTime @default(now())
@@ -34,7 +34,7 @@ model Model {
   type String
   engine String
   fuel_type String
-  year String
+  year Int
   created_at DateTime @default(now())
   updated_at DateTime @updatedAt
   deleted_at DateTime?
@@ -51,8 +51,6 @@ model Version {
   id String   @id @default(uuid())
   name String
   price String
-  features String
-  performance String
 
   model Model @relation(fields: [model_id], references: [id])
   model_id String
