+++ src/modules/BrandEntity/services/DeleteBrandService.ts
@@ -1,0 +1,9 @@
+import { BrandRepository } from "@modules/BrandEntity/repositories/BrandRepository";
+
+export default class DeleteBrandService {
+  async execute(id: string): Promise<void>{
+    const brandRepository = new BrandRepository();
+    
+    await brandRepository.delete(id)
+  }
+}
