+++ src/modules/BrandEntity/services/UpdateBrandService.ts
@@ -1,0 +1,14 @@
+import { Prisma, Brand } from "@prisma/client";
+import { BrandRepository } from "@modules/BrandEntity/repositories/BrandRepository";
+
+export default class UpdateBrandService {
+  async execute(data: Prisma.BrandUpdateInput): Promise<Brand | undefined> {
+    const brandRepository = new BrandRepository();
+
+    const brand = await brandRepository.save({
+      ...data
+    })
+
+    return brand;
+  }
+}
