+++ src/modules/ModelEntity/services/ListModelService.ts
@@ -1,0 +1,9 @@
+import { ModelRepository } from "@modules/ModelEntity/repositories/ModelRepository";
+
+export default class ListModelService {
+  async execute() {
+    const modelRepository = new ModelRepository();
+    const models = await modelRepository.find();
+    return models;
+  }
+}
