+++ src/modules/ModelEntity/services/UpdateModelService.ts
@@ -1,0 +1,14 @@
+import { Prisma, Model } from "@prisma/client";
+import { ModelRepository } from "@modules/ModelEntity/repositories/ModelRepository";
+
+export default class UpdateModelService {
+  async execute(data: Prisma.ModelCreateInput): Promise<Model| undefined> {
+    const modelRepository = new ModelRepository();
+
+    const model = await modelRepository.save({
+      ...data
+    })
+
+    return model;
+  }
+}
