+++ src/modules/VersionEntity/repositories/VersionRepository.ts
@@ -1,0 +1,50 @@
+import { Prisma, Version } from "@prisma/client";
+import prisma from "@utils/prisma";
+
+interface IRequest {
+    id?: string
+    name: string
+    price: string
+    model_id: string
+}
+
+export class VersionRepository {
+
+    async create(data: IRequest) {
+        const createVersion = await prisma.version.create({
+            data: {
+                name: data.name,
+                price: data.price,
+                model: {
+                    connect: {
+                        id: data.model_id
+                    }
+                }
+            }
+        })
+        return createVersion;
+    }
+
+    async save(version: Prisma.VersionUpdateInput): Promise<Version | undefined> {
+        const updatedVersion = await prisma.version.update({
+          where: {
+            id: (version.id as string)
+          },
+          data: version
+        })
+        return updatedVersion;
+    }
+
+    async find() {
+        const versions = prisma.version.findMany();
+        return versions;
+    }
+
+    async delete(id: string) {
+        await prisma.version.delete({
+            where: {
+                id: id
+            }
+        })
+    }
+}
