+++ src/modules/VersionEntity/services/DeleteVersionService.ts
@@ -1,0 +1,9 @@
+import { VersionRepository } from "@modules/VersionEntity/repositories/VersionRepository";
+
+export default class DeleteVersionService {
+    async execute(id: string): Promise<void>{
+      const versionRepository = new VersionRepository();
+      
+      await versionRepository.delete(id);
+    }
+  }
