+++ src/modules/VersionEntity/services/ListVersionService.ts
@@ -1,0 +1,9 @@
+import { VersionRepository } from "@modules/VersionEntity/repositories/VersionRepository";
+
+export default class ListVersionService {
+    async execute() {
+      const versionRepository = new VersionRepository();
+      const version = await versionRepository.find();
+      return version;
+    }
+  }
